Rename proj and art pages to projects and misc
The "proj" page was a shortened version of "projects" just so that it
would fit in four characters, but that is kind of dumb, so it has been
renamed "projects".

I want the "art" page to be able to host more than just art-related
things, so it has been renamed to "misc".

This change breaks links on the website, but I'm sure that ~zero people
have anything from this site bookmarked, so I think it should be fine.